chore: extract broadcast-changes suite into src/suites/broadcast-changes.ts
Part of REAL-1037.
